requirements.txt
snowflake-snowpark-python[pandas]
snowflake-cli-labs
snowflake
dotenv
google-cloud-storage